New Search Bar, Font Renderer changes, general fixes, fixed Sounds, Some setting changes and icon rendering

+++ src/main/java/dev/heliosclient/module/settings/Separator.java
@@ -15,12 +15,12 @@ public class Separator extends Setting<Boolean> {
 
     @Override
     public void render(DrawContext drawContext, int x, int y, int mouseX, int mouseY, TextRenderer textRenderer) {
-        Renderer2D.drawRectangle(drawContext.getMatrices().peek().getPositionMatrix(), x + 2, y + height / 2, width - 2, 1, 0xCCFFFFFF);
+        Renderer2D.drawRectangle(drawContext.getMatrices().peek().getPositionMatrix(), x + 2, y + (float) height / 2, width - 2, 1, 0xCCFFFFFF);
     }
 
     @Override
     public void renderCompact(DrawContext drawContext, int x, int y, int mouseX, int mouseY, TextRenderer textRenderer) {
-        Renderer2D.drawRectangle(drawContext.getMatrices().peek().getPositionMatrix(), x + 2, y + heightCompact / 2, widthCompact - 2, 1, 0xCCFFFFFF);
+        Renderer2D.drawRectangle(drawContext.getMatrices().peek().getPositionMatrix(), x + 2, y + (float) heightCompact / 2, widthCompact - 2, 1, 0xCCFFFFFF);
     }
 
 }
